RFC: start without `otk.variable` just `$`
The spec currently defines two ways to use variables. All the examples
use only a single one even in places that are technically not allowing
it. But I have to say I really like the short form this RFC advocates
to just start with that short form and see if we actually have a use
case for the long form.

The rule would be simple:
- string vars can be used inside strings "foo: hello $world" works if
  "world" is a string
- non-string vars must stand on their own, so "foo: $world" is the
  only allowed form for maps/seq etc.