RFC: simplify customizations in the tree.yaml
For a lot of the OS customizations we really only change a single
option. It would be nice to express this in a shorter way and I
tried to capture it in this commit. Essentially the customization
just moves a level deeper which AIUI is something that is already
allowed with the current way customizations are specified.