poc: initial implementation
This is the initial implementation of the Proof of Concept for the
Omnifest Toolkit (`otk`). `otk` is implemented in Python as it's the
shared language of choice between the creators and the users of
`osbuild-mpp` which `otk` aims to improve upon.

The `otk` binary is laid out as follows:

- `otk compile` compiles an omnifest into its targets.

The `otk` package is laid out as follows:

- `otk/parse` contains parsing code for full schemas and single primitives
  such as names or variables, it also contains wrapping types for
  internal types.
- `otk/help` contains helper functions.
- `otk/transform` contains the logic for transforming trees including
  the directive implementations in the directive submodule.
- `otk/error.py` contains `otk` error types, all errors inherit from
  the `OTKError` type.
- `otk/tree.py` contains types and functions to deal with the tree
  transformations that are at the core of `otk`.
- `otk/command.py` contains the subcommands as supported by the `otk` entry
  point.

Other than the above the repository also contains:

- `test/` contains unittests for `otk`.
- `doc/` contains Sphinx documentation for `otk`.
- `example/` contains YAML example files for `otk`.

Signed-off-by: Simon de Vlieger <cmdr@supakeen.com>