doc: re-import
I moved the docs out of this repository to include them in our
documentation. As this project is still in a proof of concept state it
is better to have this live in the repository.

Let's keep the docusaurus layout so it's easier to later auto-import it
into the docs website.

Signed-off-by: Simon de Vlieger <cmdr@supakeen.com>