RFC: define a single `otk.op.join`
I wonder if we can simply avoid having the user care about the type when
doing a join operation. It seems mostly a burden to the user and we need
to check during compilation of the otk file anyway so maybe we can just
leave it out?

P.S. Having an example use-case for both would be really good, I'm wonder
if we only should add things to the spec once we actually use it in the
example/* tree ?