directive: remove `otk.meta` and use `target.*` instead
This is (very much) a RFC/brainstorm commit. Right now the use
of `otk.meta` is a bit unclear, there is an open bug about this
(c.f. https://github.com/osbuild/otk/issues/2).

But it seems to me that `meta` is really about writing json data
from an omnifest for a specific tool (in this case "composer").
Which makes me think that we have support for this already with
out `target` directive. So `otk -t osbuild-composer` would generate
the needed meta information.