RFC: loops for customizations
When trying to port a slightly more elaborate customization to otk
I noticed that many import ones (like users) are actually lists
of items. Our synatx is currently not expressive enough for this.

There are (broadly) two ways to solve this:
1. run an external helper to deal with complex customizations
2. introduce a "loop" construct to the otk format

This is a stawman/draft for (2). I'm not sure how much I like it
yet, I have some sympathy for (1) but I also think it's important
to dicuss it and the loop itself does not look too bad.