otk: use filedescriptors instead of paths when reading files
The upside of using file descriptors is that there are less
TOCTOU issues, i.e. `if not src.exists()` on a path is racy
but this can easily be avoided by just opening the file and
keeping it open.

It will also make testing for stdin/stdout easier from the
test POV there is nothing special about them (i.e. it's
not even needed to test that they work just that when "-"
is passed on the commandline (or nothing) stdin/stdout is
opened).