otk: introduce traversal State
This commit adds a new `traversal.State` that is part of the `resolve()`
recursion. It is needed to keep (recursive) state about what state the
parsing is in. Currently this is used to keep track of what include is
used so that nested includes work and also so that errors can know in
what file the issue happened.