otk: rename desugar() to substitute_vars()
The name `desugar()` was in reference to initial specifications that
supported multiple ways of referring to variables and the `${name}` form
was considered "syntactic sugar".  Now we only have one way of referring
to variables and the concept of "desugaring" might be hard to
understand.

Rename it to `replace_vars()` to make its purpose clearer.

The docstring is edited to clearly separate the two scenarios we
support when replacing variables.

The exception message in case of a bad replacement value is adjusted to
print the "simple" name of the type that caused the issue.