otk: add defines and copy() method to traversal.State
The defines property of the State will be used to keep track of
subblocks of variable definitions when processing otk.define directives.

The copy() method creates a new State with replaced or copied
properties.

This will enable us to create State objects with references to
sub-objects of the global defines variables.  While processing defines
sub-objects, the State's defines can be written to to define values
nested under the key being processed, while using the global Context
object to resolve variable names inside defines.