otk: only try to merge dict with dict in defines
If a key was previously defined in the context's defines but was not a
dict and was instead a list or primitive type, we would try to merge
them, which would break when trying to index the existing value as a
dictionary.

Make sure a subblock is reused only when it's actually a dictionary.  In
all other cases, the new value replaces the old.