otk: simplify Omnifest and use `process_include` in entrypoint
With the recent change to require a fully processed tree before
creating the omnifest [0] it seems we can simplify the Omnifest
a bit more now that it essentially is just a container for an
externally resolved tree.

This commit does that by removing some of the entrypoints
like `from_yaml_{file,data}` as those are not that useful
without resolving the tree first.

One nice effect of this change is that the circular import
detection now includes the top-level entrypoint in the error
and it is also more symetrical (a single way to open/read/parse
an otk file).

The handling of data from <stdin> is not super nice right now
and it actually makes me wonder if we should maybe stop supporting
omnifests from stdin. It seems that most omnifests will contain
includes so the notion of creating an omnifest from <stdin> is
a bit fuzzy.

[0] https://github.com/osbuild/otk/pull/128