otk: disallow "otk.include" in "otk.define"
There is a bit of discussion in issue #116 about this, in a nutshell
the issue with allowing it is that it can be confusing to follow
the flow of nesting when looking at the yaml files. So for now
it might be better to not allow it. Not allowing it also simplifies
the code a bit.

Closes: https://github.com/osbuild/otk/issues/116