otk: resolve targets in a full pass
This commit changes the code so that the tree for a specific target
is resolved in a full pass. This fixes the issue that relative includes
in included targets will fail (e.g. test/data/base/13-include-in-subdir).

The commit also includes some discussion points and XXX/questions
that can be discussed/resolved later.