otk: remove unused "defines"  getter/setter from Context
It seems like nothing is currently using Context.defines plus
there is already a `Context.define()` and a `Context.variable()`
which provide similar functionality (and more as they can resolve
dotted notation). Unless there is a use-case or a good reason
let's remove these methods and bring them back if/when we need
them.