test: simplify naming of yaml/targets
We allow the use of `otk.define.foo` or `otk.target.osbuild.foo`
to ensure when multiple `otk.defines` are done in a single file
it is still valid yaml. However in a lot of our examples we only
have a single `otk.define` in the file or only a single target.
Here there is no need to append a postfix.

This commit simplifies these occurances and removes unneeded
postfixes. They are still valid and some tests needs them but
the are no longer there by default.

We should probably do the same for the `examples/**`.