otk: check for targetless keys only when resolving without target
The api/code here is confusing, we want to make sure that we do
not allow "values" in the tree that have no targets. But this
is now tricky when we resolve with a target. So we are currently
forced to resolve without a target for error checking and then
to do the real resolving. But it would be nice to not require a
dry-run (targetless) resolve. This commit adds a workaround but
this needs some proper thinking and refactoring.