test: add tests for directives that do not allow siblings
It is not allowed to have multiple `otk.op` or `otk.external` in
the same dict today, add a test that makes this explicit.

Fwiw, it's not clear to me why we disallow this but given that
it seems a concious decision let's ensure we not change it by
accident.