otk: fix incorrect tree copy in `otk.include`
When trying to compile the example in PR#167 I noticed slightly
strange behavior. Currently when doing:
```
otk.target.osbuild:
  otk.include: "16-include-copy/repository-${version}.yaml"
```
the output is
```
{
     'otk.include': '16-include-copy/repository-${version}.yaml',
     'sources': {},
     'version': '2',
}
```
The reason is that we have an extra `tree.copy()` when we process
includes. However in other parts of the code (including accross
recursions) we use "tree.update()" so that the tree we worked
on and the tree that the other parts of the recursion go out of
sync. The easiest fix is to simply drop the copy.