otk: error when a `otk.include` inside a dict override the dict
We allow the following:
```yaml
otk.target.osbuild:
  sub:
    c: d
    otk.include: "foo.yaml"
```
The current semantic is that if the resolved value of `foo.yaml` is in
itself a dict the two are merged and if it's a non-dict value
the `sub` tree is replaced with the scalar value from `foo.yaml`.

However in the above example this can lead to unexpected results,
i.e. depending on how `foo.yaml` resolves `c: d` is removed or
kept.

This commit make it an error if an `otk.include` inside a non
empty dict returns a scalar. We *could* consider also making it
a warning but I fail to see a use-case right now (that is not
contrived) where this would not be unexpected/unwanted behavior.