[RFC] test: test against reference images lib generated manifest
This is an idea how to make development of the otk manifest
easier to iterate on. The idea is that we filter out the
sources/rpms (we should consider later to keep the rpm names but
not the exact versions) and then compare the stages. This way
we can iterate on the development of the centos-9 example and
always be confident that the reference is matching.

The reference was generated with the `images` library.