test: use `sort_keys=False` in the `gen-image-def` helper
When generating the "images" library reference yaml we used the
default yaml.dump() implementation. This will reoder the keys
which is undesirable for us as it moves e.g. the `type: ...`
key at the end of a block. But in a osbuild manifest this is
the one of the most important pieces of information.

So this commit changes the code to not sort the yaml keys.