test: workaround different RNG results in `gen-manifests`
This is a hacky commit - the OSBUILD_TESTING_RNG_SEED is set for both
"gen-manifests" and "otk-gen-partition-tables" but the partition
table code is called slightly differently which means that the UUIDs
get out of sync. We need this (f)ugly mapping here to fix that.

Fortunately both sides are deterministic so this mapping is hopefully
enough.