Makefile: point "images" to "@main" to get HEAD
The `go install repo@latest` means that go install will fetch
the latest released version of the repository. However with
the speed that we develop we need to get the latest HEAD insead,
this can be expressed via "go install repo@main". This commit
uses that now so that we can get the brand-new
`otk-make-grub2-inst-stage`.