test: add a new `normalize_rpm_refs` helper
This commit adds a new `normalize_rpm_refs()` helper that will
normalize the rpm file inputs for comparison, i.e. it will sort
the hashes and remove duplicates. This is important because the
"images" library writes duplicated input rpm packages from the
inputs as duplicated hashes into the references.

But in our test we only need to worry about that exactly the right
rpm files got included, the order is not important.