gen-depsolve-dnf4: expose only `const.kernel.{name,version}`
When doing a `gen-depsolve-dnf4` export only the kernel name and
version in the result set instead of every package name with the
version. Mainly to limit the API surface we need to support, for
now we know we need to handle the kernel special but it's unclear
if we need to expose details about the other resolved packages.