test: tweak gen-images() to be smarter about the UUID hacks
This commit tweaks the gen-images-ref to be slightly smarter
about how to handle UUIDs. Instead of the extremly naive table
of uuids generated by "images" and "gen-partition-table" it now
tries to figure out the UUID from images via the "mkfs" stage
of the manifest. This is also a bit (f)ugly but avoids some
busywork.

Hopefully we find a better way but this should give us some
time.