external: add new `get-dnf4-package-info` external
This commit adds a new `get-dnf4-package-info` external that can
be used to extract details about a dnf4 resolved package.

The main use-case is the kernel package where we need some details
like the version number to put into the grub stage. With this
commit we can do something like:
```yaml
  kernel:
    package:
      otk.external.osbuild-get-dnf4-package-info:
        packageset: ${packages.os}
        packagename: ${modifications.kernel_package_name}
```