example/centos: extract image pipeline for luks-lvm images
The image pipeline for other image types contains a call to the
partition stage generators.  However, the partition table generator does
not yet support LUKS containers, so we can't use it to generate the
partition table for the edge-raw-image.

Another alternative would be to import the full description of the
partition table as it would be generated when we get support, by
serialising the base PT of the edge-raw-image.  This would allow us to
use the stage generators of the standard image pipeline with a
hard-coded LUKS+LVM partition table description.  However, that requires
defining:

```
otk.define:
  filesystem:
    tree:
      const:
        filename: disk.img
        internal:
          partition-table:
            ...
```

and the key `partition-table` is not valid in otk.

Add note to the top of the file about LUKS support requirement.