example/centos: symlink OCI
OCI stands for Oracle Cloud Infrastructure. For hysterical raisins this
is a separate image type that is identical to the QCOW2 image type for
the same architecture.

Let's import the reference omnifest but directly symlink the file to the
QCOW2 omnifest.

The hysterical raisin is that making OCI and alias of QCOW2 in `images`
or `osbuild-composer` might make it not show up in (some) APIs.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>