example/centos: resolve ostree commit for edge-raw-image
Use the otk-resolve-ostree-commit external to resolve the commit ID for
the edge-raw-image.

Replace all instances of the ref and URL with the values from the
external.

Note that the input references to the org.osbuild.ostree.pull stage need
to be objects keyed by the commit ID, which is a variable.  We don't
support using variables for object keys (yet).  Added a note in the yaml
file to indicate this.