pre-commit-config.yaml: drop `id: check-yaml` (cargo culted?)
This commit drops the `id: check-yaml` from the pre-commit hook.
We already have a dedicated `yamllint` hook and for the tests
we have some (intentionally) invalid yaml files that need to
be ignored.

I wasted^Wspend 30min trying to figure out how to disable
the test files with `check-yaml` but couldn't figure it
out and given that we already have a linter it seems overkill
to have two that need to be confiured separately.