test: add test of customizations against reference images
This commit extends the testing of otk files to also support
checking that our customizations are correctly covered. This
is done via a new reference image type with full customizations.

For all refrence images that have this full customizatons we
generate a:
```yaml
otk.define:
 user:
  modifications:
   locale: nl_NL.UTF-8
   ...
```
fragment and compare out customizations with the ones that `images`
generated.