otk/command: `-V/--version`
Implement a `-V/--version` option to the root command. Note that this
changes the behaviour on using `otk` without a subcommand given.

To have `otk -V` work the subcommand needs to be made optional so the
root entrypoint has been changed to print usage if no subcommand was
given.

This means that the test for the parser without a subcommand also gets
dropped as the behavior is no longer an error case.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>